+++ agent_pool.go
@@ -72,6 +72,9 @@ type AgentPoolListOptions struct {
 	// Optional: A list of relations to include. See available resources
 	// https://www.terraform.io/cloud-docs/api-docs/agents#available-related-resources
 	Include []AgentPoolIncludeOpt `url:"include,omitempty"`
+
+	// Optional: A search query string used to filter agent pool. Agent pools are searchable by name
+	Query string `url:"q,omitempty"`
 }
 
 // AgentPoolCreateOptions represents the options for creating an agent pool.